requirements.txt
openai
python-dotenv
pyyaml
termcolor
